Move `batch_skip_reason` to `dotpkg.manifest`
